requirements.txt
yacs==0.1.8
numpy==1.21.2
torchvision==0.8.2+cu110
tqdm==4.62.3
opencv_contrib_python==4.5.3.56
imageio==2.9.0
torch==1.7.1+cu110
mmcv_full==1.3.10
matplotlib==3.4.3
mmcv==1.3.17
Pillow==8.4.0
scikit_learn==1.0.1

